Added Autocomplete trait with generic complete function and use it on all commands
