+++ src/commands/create-production-site.php
@@ -11,6 +11,8 @@ use Symfony\Component\Console\Helper\ProgressBar;
 use function Team51\Helper\run_app_command;
 
 class Create_Production_Site extends Command {
+	use \Team51\Helper\Autocomplete;
+
 	protected static $defaultName = 'create-production-site';
 
 	const DEPLOYHQ_ZONE_EUROPE  = 3; // UK
