+++ src/commands/pressable-site-rotate-wp-user-password.php
@@ -36,6 +36,8 @@ use function Team51\Helper\update_1password_item;
  * CLI command for rotating the WP password of users on Pressable sites.
  */
 final class Pressable_Site_Rotate_WP_User_Password extends Command {
+	use \Team51\Helper\Autocomplete;
+
 	// region FIELDS AND CONSTANTS
 
 	/**
