+++ src/commands/pressable-site-run-wp-cli-command.php
@@ -18,6 +18,8 @@ use function Team51\Helper\maybe_define_console_verbosity;
  * CLI command for running a WP-CLI command on a Pressable site.
  */
 final class Pressable_Site_Run_WP_CLI_Command extends Command {
+	use \Team51\Helper\Autocomplete;
+
 	// region FIELDS AND CONSTANTS
 
 	/**
